Fixed project files
Changed project from Project 5 to Project 6. The class I was in
required all of these projects to be named Project n, and I realized
that project 6 is the code comparison tool.
Note that I’ve implemented some third party libraries to do the string
comparison algorithms here, so that I did not have to design my own
string comparison algorithm.